requirements.txt
fastapi
uvicorn[standard]
pymongo
groq
python-dotenv
